requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
bcrypt==4.2.0
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
yarl==1.17.1
